requirements.txt
PuLP==2.8.0
matplotlib==3.8.4
matplotlib-inline==0.1.7
seaborn==0.13.2
notebook==7.1.3
pandas==2.2.2
PyYAML==6.0.1